#54 - Feat : 거래내역 상세보기 구현 완료

+++ BE/src/main/java/com/ssafy/boney/domain/transaction/dto/TransactionResponseDto.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class TransactionResponseDto {
-    private Long transactionId;
+    private Integer transactionId;
     private LocalDateTime transactionDate;
     private String transactionContent;
     private Long transactionAmount;
